move caching/dir delete code in own files
